Breaking update: fixed determinism in multiple places, removed token whitelist, replaces multiple transaction base types with transaction and commitment types, added separate gas limit for commitment transactions, merged attestation transaction into depository attestation, removed delegation/consensus/attestation transactions, improved erd64 proposer selection alg, oracle will now use only inclusive fees for withdrawals

+++ src/tangent/kernel/cell.h
@@ -828,7 +828,7 @@ namespace vitex
 		template <>
 		struct key_hasher<tangent::cell::string_repr>
 		{
-			typedef float argument_type;
+			typedef int argument_type;
 			typedef size_t result_type;
 			using is_transparent = void;
